Add progress cleanup on cancel option. Cancelled downloaded files are deleted but tiles from tile generation are not.

+++ src/mil/nga/giat/geopackage/io/GeoPackageIOUtils.java
@@ -126,7 +126,7 @@ public class GeoPackageIOUtils {
 		copyStream(copyFrom, to, progress);
 
 		// Try to delete the file if progress was cancelled
-		if (progress != null && !progress.isActive()) {
+		if (progress != null && !progress.isActive() && progress.cleanupOnCancel()) {
 			copyTo.delete();
 		}
 	}
